Style: Manage "My Site" items margins using the newly added decoration class
"MySiteCardAndItemDecoration" is now used to set margins for both cards and items on the "My Site" tab.
