+++ WordPress/src/main/res/layout/debug_settings_header.xml
@@ -3,7 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:minHeight="@dimen/menu_item_height">
+    android:minHeight="@dimen/menu_item_height"
+    android:paddingStart="@dimen/margin_extra_large"
+    android:paddingEnd="@dimen/margin_extra_large">
 
     <com.google.android.material.textview.MaterialTextView
         android:gravity="start"
