+++ WordPress/src/main/res/values/styles.xml
@@ -522,8 +522,6 @@
     <style name="MySiteListRowLayout">
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingEnd">@dimen/content_margin_site_row_end</item>
-        <item name="android:paddingStart">@dimen/content_margin_site_row_start</item>
         <item name="android:foreground">?android:attr/selectableItemBackground</item>
     </style>
 
@@ -625,7 +623,6 @@
         <item name="android:layout_height">wrap_content</item>
         <item name="android:layout_width">match_parent</item>
         <item name="android:paddingBottom">@dimen/my_site_margin_general</item>
-        <item name="android:paddingStart">@dimen/content_margin_site_row_start</item>
         <item name="android:paddingTop">@dimen/my_site_margin_general</item>
         <item name="android:textColor">?attr/wpColorOnSurfaceMedium</item>
         <item name="android:textAppearance">?attr/textAppearanceSubtitle2</item>
@@ -1232,6 +1229,8 @@
     <!-- Main actions styles -->
     <style name="MainBottomSheetRowLayout" parent="MySiteListRowLayout">
         <item name="android:minHeight">@dimen/main_bottom_sheet_list_row_min_height</item>
+        <item name="android:paddingEnd">@dimen/content_margin_site_row_end</item>
+        <item name="android:paddingStart">@dimen/content_margin_site_row_start</item>
     </style>
 
     <style name="MainBottomSheetRowIcon" parent="MySiteListRowIcon">
